[Bug Fix, Code Cleanup, Optimizations]
Bug: Fixed infinite scrobbling of video / song pair.
Cleanup: Minimized duplicate code, replaced try except with get default
Optimization: Changed what is stored in the database to track_id only, removed unused columns.

To be addressed: Existing instances of the old version will require data.db to be removed prior to using the updated source.

Possible solution: Check if the table matches the new structure, if not, drop the database and restructure. [May lead to duplicate scrobbles if the user had listened to songs that day]